requirements.txt
jupyterlab
langchain
langchain-experimental
openai
psycopg2-binary
tiktoken
faiss-cpu
ipython
matplotlib
streamlit